shared/pyproject.toml
[tool.poetry]
name = "sentilyze-core"
version = "4.0.0"
description = "Shared core library for Sentilyze services"
authors = ["Sentilyze Team <team@sentilyze.live>"]
packages = [{ include = "sentilyze_core" }]

[tool.poetry.dependencies]
python = "^3.11"

# Config / models
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Logging
structlog = "^24.1.0"
python-json-logger = "^2.0.0"

# Retry/backoff
tenacity = "^8.2.0"

# Cache clients
redis = "^5.0.0"

# GCP clients
google-api-core = "^2.15.0"
google-cloud-bigquery = "^3.15.0"
google-cloud-pubsub = "^2.21.0"
google-cloud-firestore = "^2.14.0"
google-cloud-secret-manager = "^2.18.0"

# Metrics
prometheus-client = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

